Add fastfetch configuration and improve installation script and README

+++ install.sh
@@ -72,18 +72,30 @@ if [ ! -d "$HOME/.oh-my-zsh" ]; then
 fi
 
 # Install Powerlevel10k theme for Oh My Zsh
-git clone --depth=1 https://github.com/romkatv/powerlevel10k.git ${ZSH_CUSTOM:-$HOME/.oh-my-zsh/custom}/themes/powerlevel10k
+if [ ! -d "${ZSH_CUSTOM:-$HOME/.oh-my-zsh/custom}/themes/powerlevel10k" ]; then
+  git clone --depth=1 https://github.com/romkatv/powerlevel10k.git ${ZSH_CUSTOM:-$HOME/.oh-my-zsh/custom}/themes/powerlevel10k
+else
+  echo "Powerlevel10k already installed, skipping clone..."
+fi
 
 # Install zsh plugins
-git clone https://github.com/zsh-users/zsh-autosuggestions ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-autosuggestions
-git clone https://github.com/zsh-users/zsh-syntax-highlighting.git ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-syntax-highlighting
+if [ ! -d "${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-autosuggestions" ]; then
+  git clone https://github.com/zsh-users/zsh-autosuggestions ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-autosuggestions
+fi
+if [ ! -d "${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-syntax-highlighting" ]; then
+  git clone https://github.com/zsh-users/zsh-syntax-highlighting.git ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-syntax-highlighting
+fi
 
 # Install Powerline Fonts
-git clone https://github.com/powerline/fonts.git --depth=1
-cd fonts
-./install.sh
-cd ..
-rm -rf fonts
+if ! fc-list | grep -i "Powerline" > /dev/null; then
+  git clone https://github.com/powerline/fonts.git --depth=1
+  cd fonts
+  ./install.sh
+  cd ..
+  rm -rf fonts
+else
+  echo "Powerline fonts already installed, skipping..."
+fi
 
 # Get the directory where this script is located
 DOTFILES_DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
@@ -96,6 +108,11 @@ cp -f "$DOTFILES_DIR/.gitconfig" $HOME/.gitconfig
 echo "Copying .zshrc and .p10k.zsh to home directory..."
 cp -f "$DOTFILES_DIR/.zshrc" $HOME/.zshrc
 cp -f "$DOTFILES_DIR/.p10k.zsh" $HOME/.p10k.zsh
+
+# Copy fastfetch config
+echo "Copying fastfetch config..."
+mkdir -p "$HOME/.config/fastfetch"
+cp -f "$DOTFILES_DIR/fastfetch/config.jsonc" "$HOME/.config/fastfetch/config.jsonc"
 
 # Set macOS preferences - we will run this last because this will reload the shell
 # source ./.macos
